make unwrap_test_universe a class UnWrapUniverse in MDAnalysisTests.core.util
